Unthought/unthought/requirements.txt
appdirs==1.4.3
asgiref==3.3.4
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.0
channels==3.0.3
chardet==3.0.4
comtypes==1.1.7
constantly==15.1.0
cryptography==3.4.7
cycler==0.10.0
daphne==3.0.2
decorator==4.4.2
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.2.1
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.0
docopt==0.6.2
docutils==0.16
ez-setup==0.9
filelock==3.0.12
gunicorn==20.0.4
hyperlink==21.0.0
idna==2.10
imageio==2.9.0
incremental==21.3.0
jedi==0.15.2
Js2Py==0.70
kivy==2.0.0rc1
kivy-deps.angle==0.1.10
kivy-deps.glew==0.1.12
kivy-deps.gstreamer==0.1.18
kivy-deps.sdl2==0.1.23
kivy-garden==0.1.4
kiwisolver==1.3.1
matplotlib==3.3.3
more-itertools==8.6.0
networkx==2.5
numpy==1.19.4
opencv-python==4.2.0.32
packaging==20.8
parso==0.7.0
Pillow==8.0.1
pipwin==0.5.0
pluggy==0.13.1
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/HP/pipwin/PyAudio-0.2.11-cp38-cp38-win32.whl
pycparser==2.20
pygame==1.9.6
pygments==2.6.1
pyjsparser==2.7.1
PyMySQL==0.9.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypiwin32==223
PyPrind==2.11.2
pySmartDL==1.3.4
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
pyttsx3==2.90
pytz==2019.3
pywin32==227
requests==2.23.0
selenium==3.141.0
service-identity==21.1.0
six==1.15.0
soupsieve==2.1
SpeechRecognition==3.8.1
sqlparse==0.3.1
sudo==1.0.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
tzlocal==2.1
urllib3==1.25.8
virtualenv==20.0.18
whitenoise==5.1.0
wikipedia==1.4.0
wolframalpha==4.1.1
xmltodict==0.12.0
zope.interface==5.4.0
